Rts/elf-linker: Upcast to 64bit to satisfy format string.
The elf size is 32bit on 32bit builds and 64 otherwise.
We just upcast to 64bits before printing now.

+++ rts/linker/Elf.c
@@ -904,8 +904,8 @@ ocGetNames_ELF ( ObjectCode* oc )
                    ASSERT(common_used <= common_size);
 
                    IF_DEBUG(linker,
-                            debugBelch("COMMON symbol, size %lu name %s allocated at %p\n",
-                                       symbol->elf_sym->st_size, nm, symbol->addr));
+                            debugBelch("COMMON symbol, size %llu name %s allocated at %p\n",
+                                       (uint64_t) symbol->elf_sym->st_size, nm, symbol->addr));
 
                    /* Pointless to do addProddableBlock() for this area,
                       since the linker should never poke around in it. */
